providers/requirements.txt
kafka==1.3.5
python-binance~=1.0.15
websocket-client~=1.2.3
